Devided Transaction into Deposit, Withdraw and Transfer

+++ users/models.py
@@ -6,18 +6,18 @@ from django.db import models
 class Club(models.Model):
     admin = models.OneToOneField('User', on_delete=models.SET_NULL, null=True, help_text="Club admin id")
     name = models.CharField(max_length=255, help_text="Name of the club")
-    balance = models.DecimalField(max_digits=15, decimal_places=2, default=0,
-                                  validators=[MinValueValidator(0)],
-                                  help_text='Club\'s current balance.')
+    balance = models.FloatField(default=0,
+                                validators=[MinValueValidator(0)],
+                                help_text='Club\'s current balance.')
 
     def __str__(self):
         return self.name
 
 
 class User(AbstractUser):
-    balance = models.DecimalField(max_digits=15, decimal_places=2, default=0,
-                                  validators=[MinValueValidator(0)],
-                                  help_text='Users current balance. ')
+    balance = models.FloatField(default=0,
+                                validators=[MinValueValidator(0)],
+                                help_text='Users current balance. ')
     user_club = models.ForeignKey(Club, on_delete=models.SET_NULL, null=True, help_text='Users current club.')
     email = models.EmailField('Email address', unique=True)
     phone = models.CharField('Phone number', max_length=18, unique=True)
